requirements.txt
aniso8601==3.0.0
click==6.7
Flask==1.0.2
Flask-JWT-Simple==0.0.3
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.8.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
marshmallow==2.15.2
PyJWT==1.6.1
pytz==2018.4
six==1.11.0
SQLAlchemy==1.2.7
Werkzeug==0.14.1
